Put in LeagueEntryDTO class to LeaguePosition class
